#260 - initial work
