+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/SourceMentor.kt
@@ -45,6 +45,15 @@ class SourceMentor : CoroutineVerticle() {
         }
     }
 
+    override suspend fun stop() {
+        running = false
+        stillValidTasks.clear()
+        taskQueue.clear()
+        jobList.clear()
+        adviceListeners.clear()
+        log.info("SourceMentor stopped")
+    }
+
     private suspend fun runJobProcessing() {
         running = true
         while (running) {
@@ -85,6 +94,7 @@ class SourceMentor : CoroutineVerticle() {
     }
 
     private suspend fun executeTask(job: MentorJob, task: MentorTask) {
+        if (!running) return
         job.log("Executing task: $task")
         task.executeTask(job)
         if (!task.asyncTask) {
@@ -109,6 +119,7 @@ class SourceMentor : CoroutineVerticle() {
         jobsWhichRequireTask: List<MentorJob>,
         currentTask: MentorTask
     ) {
+        if (!running) return
         val tasksStillRequired = Collections.newSetFromMap(
             IdentityHashMap<MentorTask, Boolean>()
         )
