+++ plugin/jetbrains/build.gradle.kts
@@ -33,6 +33,9 @@ intellij {
     setPlugins("java", "Groovy", "Kotlin", "org.intellij.scala:2020.2.27")
 }
 tasks.getByName("buildSearchableOptions").onlyIf { false } //todo: figure out how to remove
+tasks.getByName<JavaExec>("runIde") {
+    systemProperty("sourcemarker.debug.capture_logs", true)
+}
 
 repositories {
     maven(url = "https://jitpack.io") { name = "jitpack" }
