+++ plugin/jetbrains/src/main/kotlin/com/sourceplusplus/sourcemarker/SourceMarkerPlugin.kt
@@ -1,0 +1,321 @@
+package com.sourceplusplus.sourcemarker
+
+import com.fasterxml.jackson.core.JsonGenerator
+import com.fasterxml.jackson.core.JsonParser
+import com.fasterxml.jackson.databind.*
+import com.fasterxml.jackson.databind.module.SimpleModule
+import com.fasterxml.jackson.datatype.guava.GuavaModule
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import com.intellij.ide.ui.laf.IntelliJLaf
+import com.intellij.ide.util.PropertiesComponent
+import com.intellij.notification.Notification
+import com.intellij.notification.NotificationType
+import com.intellij.notification.Notifications
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.project.Project
+import com.intellij.psi.JavaPsiFacade
+import com.intellij.psi.search.ProjectScope
+import com.sourceplusplus.marker.SourceMarker
+import com.sourceplusplus.marker.source.mark.api.component.api.config.ComponentSizeEvaluator
+import com.sourceplusplus.marker.source.mark.api.component.api.config.SourceMarkComponentConfiguration
+import com.sourceplusplus.marker.source.mark.api.component.jcef.SourceMarkSingleJcefComponentProvider
+import com.sourceplusplus.marker.source.mark.api.component.jcef.config.BrowserLoadingListener
+import com.sourceplusplus.marker.source.mark.api.component.jcef.config.SourceMarkJcefComponentConfiguration
+import com.sourceplusplus.marker.source.mark.api.filter.CreateSourceMarkFilter
+import com.sourceplusplus.marker.source.mark.gutter.config.GutterMarkConfiguration
+import com.sourceplusplus.mentor.SourceMentor
+import com.sourceplusplus.mentor.base.MentorJobConfig
+import com.sourceplusplus.mentor.impl.job.ActiveExceptionMentor
+import com.sourceplusplus.mentor.impl.job.RampDetectionMentor
+import com.sourceplusplus.monitor.skywalking.SkywalkingMonitor
+import com.sourceplusplus.portal.SourcePortal
+import com.sourceplusplus.portal.backend.PortalServer
+import com.sourceplusplus.protocol.artifact.endpoint.EndpointResult
+import com.sourceplusplus.protocol.artifact.metrics.ArtifactMetricResult
+import com.sourceplusplus.protocol.artifact.trace.TraceResult
+import com.sourceplusplus.protocol.artifact.trace.TraceSpanStackQueryResult
+import com.sourceplusplus.sourcemarker.listeners.ArtifactAdviceListener
+import com.sourceplusplus.sourcemarker.listeners.PluginSourceMarkEventListener
+import com.sourceplusplus.sourcemarker.listeners.PortalEventListener
+import com.sourceplusplus.sourcemarker.psi.PluginSqlProducerSearch
+import com.sourceplusplus.sourcemarker.settings.SourceMarkerConfig
+import io.vertx.core.Promise
+import io.vertx.core.Vertx
+import io.vertx.core.buffer.Buffer
+import io.vertx.core.eventbus.MessageCodec
+import io.vertx.core.json.Json
+import io.vertx.core.json.jackson.DatabindCodec
+import io.vertx.ext.bridge.PermittedOptions
+import io.vertx.ext.web.Router
+import io.vertx.ext.web.handler.sockjs.SockJSBridgeOptions
+import io.vertx.ext.web.handler.sockjs.SockJSHandler
+import io.vertx.kotlin.core.deployVerticleAwait
+import io.vertx.kotlin.core.http.listenAwait
+import io.vertx.kotlin.core.undeployAwait
+import io.vertx.kotlin.coroutines.dispatcher
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import kotlinx.datetime.Instant
+import org.slf4j.LoggerFactory
+import java.awt.Dimension
+import java.util.*
+import javax.swing.UIManager
+
+/**
+ * todo: description.
+ *
+ * @since 0.0.1
+ * @author [Brandon Fergerson](mailto:bfergerson@apache.org)
+ */
+object SourceMarkerPlugin {
+
+    private val log = LoggerFactory.getLogger(SourceMarkerPlugin::class.java)
+    private val deploymentIds = mutableListOf<String>()
+    val vertx: Vertx = Vertx.vertx()
+
+    init {
+        log.debug("Registering SourceMarker protocol codecs")
+        vertx.eventBus().registerDefaultCodec(SourcePortal::class.java, LocalMessageCodec())
+        vertx.eventBus().registerDefaultCodec(ArtifactMetricResult::class.java, LocalMessageCodec())
+        vertx.eventBus().registerDefaultCodec(TraceResult::class.java, LocalMessageCodec())
+        vertx.eventBus().registerDefaultCodec(TraceSpanStackQueryResult::class.java, LocalMessageCodec())
+        vertx.eventBus().registerDefaultCodec(EndpointResult::class.java, LocalMessageCodec())
+
+        val module = SimpleModule()
+        module.addSerializer(Instant::class.java, KSerializers.KotlinInstantSerializer())
+        module.addDeserializer(Instant::class.java, KSerializers.KotlinInstantDeserializer())
+        DatabindCodec.mapper().registerModule(module)
+
+        DatabindCodec.mapper().registerModule(GuavaModule())
+        DatabindCodec.mapper().registerModule(Jdk8Module())
+        DatabindCodec.mapper().registerModule(JavaTimeModule())
+        DatabindCodec.mapper().enable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING)
+        DatabindCodec.mapper().enable(DeserializationFeature.READ_ENUMS_USING_TO_STRING)
+    }
+
+    suspend fun init(project: Project) {
+        val config = if (PropertiesComponent.getInstance().isValueSet("sourcemarker_plugin_config")) {
+            Json.decodeValue(
+                PropertiesComponent.getInstance().getValue("sourcemarker_plugin_config"),
+                SourceMarkerConfig::class.java
+            )
+        } else {
+            SourceMarkerConfig()
+        }
+
+        //attempt to determine root source package automatically (if necessary)
+        val checkRootPackage = Promise.promise<Nothing>()
+        if (config.rootSourcePackage == null) {
+            ApplicationManager.getApplication().runReadAction {
+                var basePackages = JavaPsiFacade.getInstance(project).findPackage("")
+                    ?.getSubPackages(ProjectScope.getProjectScope(project))
+
+                //remove non-code packages
+                basePackages = basePackages!!.filter { it.qualifiedName != "asciidoc" }.toTypedArray()
+
+                //determine deepest common source package
+                if (basePackages.isNotEmpty()) {
+                    var rootPackage: String? = null
+                    while (basePackages!!.size == 1) {
+                        rootPackage = basePackages[0]!!.qualifiedName
+                        basePackages = basePackages[0]!!.getSubPackages(ProjectScope.getProjectScope(project))
+                    }
+                    if (rootPackage != null) {
+                        config.rootSourcePackage = rootPackage
+                        PropertiesComponent.getInstance().setValue("sourcemarker_plugin_config", Json.encode(config))
+                    }
+                }
+                checkRootPackage.complete()
+            }
+        } else {
+            checkRootPackage.complete()
+        }
+
+        checkRootPackage.future().onComplete {
+            GlobalScope.launch(vertx.dispatcher()) {
+                var connectedMonitor = false
+                try {
+                    initMonitor(config)
+                    connectedMonitor = true
+                } catch (throwable: Throwable) {
+                    //todo: if first time bring up config panel automatically instead of notification
+                    Notifications.Bus.notify(
+                        Notification(
+                            "SourceMarker", "Connection Failed",
+                            "SourceMarker failed to connect to Apache SkyWalking. " +
+                                    "Please ensure Apache SkyWalking is running and the correct configuration " +
+                                    "is set at: Settings -> Tools -> SourceMarker",
+                            NotificationType.ERROR
+                        )
+                    )
+                    log.error("Connection failed", throwable)
+                }
+
+                if (connectedMonitor) {
+                    initPortal()
+                    initMarker(config)
+                    initMapper()
+                    initMentor(config)
+                }
+            }
+        }
+    }
+
+    suspend fun restart(project: Project) {
+        if (SourceMarker.enabled) {
+            SourceMarker.clearAvailableSourceFileMarkers()
+        }
+        deploymentIds.forEach {
+            vertx.undeployAwait(it)
+        }
+        deploymentIds.clear()
+        init(project)
+    }
+
+    private suspend fun initMonitor(config: SourceMarkerConfig) {
+        deploymentIds.add(vertx.deployVerticleAwait(SkywalkingMonitor(config.skywalkingOapUrl)))
+    }
+
+    private fun initMapper() {
+        //todo: this
+    }
+
+    /**
+     * Schedules long running, generic, and low-priority mentor jobs.
+     * High-priority, specific, and short running mentor jobs are executed during source code navigation.
+     */
+    private suspend fun initMentor(config: SourceMarkerConfig): SourceMentor {
+        val mentor = SourceMentor()
+        val mentorAdviceListener = ArtifactAdviceListener()
+        SourceMarker.addGlobalSourceMarkEventListener(mentorAdviceListener)
+        mentor.addAdviceListener(mentorAdviceListener)
+
+        //configure and add long running low-priority mentor jobs
+        mentor.addJobs(
+            RampDetectionMentor(
+                vertx, PluginSqlProducerSearch()
+            ).withConfig(
+                MentorJobConfig(
+                    repeatForever = true,
+                    repeatDelay = 30_000
+                    //todo: configurable schedule
+                )
+            )
+        )
+        if (config.rootSourcePackage != null) {
+            mentor.addJob(
+                ActiveExceptionMentor(
+                    vertx, config.rootSourcePackage!!
+                ).withConfig(
+                    MentorJobConfig(
+                        repeatForever = true,
+                        repeatDelay = 30_000
+                        //todo: configurable schedule
+                    )
+                )
+            )
+        } else {
+            log.warn("Could not determine root source package. Skipped adding ActiveExceptionMentor...")
+        }
+
+        deploymentIds.add(vertx.deployVerticleAwait(mentor))
+        return mentor
+    }
+
+    private suspend fun initPortal() {
+        //todo: load portal config (custom themes, etc)
+        deploymentIds.add(vertx.deployVerticleAwait(PortalServer()))
+        deploymentIds.add(vertx.deployVerticleAwait(PortalEventListener()))
+
+        //todo: portal should be connected to event bus without bridge
+        val sockJSHandler = SockJSHandler.create(vertx)
+        val portalBridgeOptions = SockJSBridgeOptions()
+            .addInboundPermitted(PermittedOptions().setAddressRegex(".+"))
+            .addOutboundPermitted(PermittedOptions().setAddressRegex(".+"))
+        sockJSHandler.bridge(portalBridgeOptions)
+
+        val router = Router.router(vertx)
+        router.route("/eventbus/*").handler(sockJSHandler)
+        vertx.createHttpServer().requestHandler(router).listenAwait(8888, "localhost")
+    }
+
+    private fun initMarker(config: SourceMarkerConfig) {
+        SourceMarker.addGlobalSourceMarkEventListener(PluginSourceMarkEventListener())
+
+        val gutterMarkConfig = GutterMarkConfiguration()
+        gutterMarkConfig.activateOnMouseHover = false
+        gutterMarkConfig.activateOnKeyboardShortcut = true
+        val componentProvider = SourceMarkSingleJcefComponentProvider().apply {
+            defaultConfiguration.preloadJcefBrowser = false
+            defaultConfiguration.componentSizeEvaluator = object : ComponentSizeEvaluator() {
+                override fun getDynamicSize(
+                    editor: Editor,
+                    configuration: SourceMarkComponentConfiguration
+                ): Dimension {
+                    var portalWidth = (editor.contentComponent.width * 0.8).toInt()
+                    if (portalWidth > 775) {
+                        portalWidth = 775
+                    }
+                    return Dimension(portalWidth, 250)
+                }
+            }
+            defaultConfiguration.browserLoadingListener = object : BrowserLoadingListener() {
+                override fun beforeBrowserCreated(configuration: SourceMarkJcefComponentConfiguration) {
+                    val initialPortal = SourcePortal.getPortals()[0]
+                    val page = "${initialPortal.currentTab.name.toLowerCase()}.html"
+                    val darkMode = UIManager.getLookAndFeel() !is IntelliJLaf
+                    configuration.initialUrl =
+                        "http://localhost:8080/$page?portalUuid=${initialPortal.portalUuid}&dark_mode=$darkMode"
+                }
+            }
+        }
+        gutterMarkConfig.componentProvider = componentProvider
+
+        SourceMarker.configuration.defaultGutterMarkConfiguration = gutterMarkConfig
+        SourceMarker.configuration.defaultInlayMarkConfiguration.componentProvider = componentProvider
+
+        if (config.rootSourcePackage != null) {
+            SourceMarker.configuration.createSourceMarkFilter = CreateSourceMarkFilter { artifactQualifiedName ->
+                artifactQualifiedName.startsWith(config.rootSourcePackage!!)
+            }
+        } else {
+            log.warn("Could not determine root source package. Skipped adding create source mark filter...")
+        }
+    }
+
+    /**
+     * Used to transmit protocol messages.
+     *
+     * @since 0.0.1
+     */
+    class LocalMessageCodec<T> : MessageCodec<T, T> {
+        override fun encodeToWire(buffer: Buffer, o: T): Unit =
+            throw UnsupportedOperationException("Not supported yet.")
+
+        override fun decodeFromWire(pos: Int, buffer: Buffer): T =
+            throw UnsupportedOperationException("Not supported yet.")
+
+        override fun transform(o: T): T = o
+        override fun name(): String = UUID.randomUUID().toString()
+        override fun systemCodecID(): Byte = -1
+    }
+
+    /**
+     * Used to serialize/deserialize Kotlin classes.
+     *
+     * @since 0.0.1
+     */
+    class KSerializers {
+        class KotlinInstantSerializer : JsonSerializer<Instant>() {
+            override fun serialize(value: Instant, jgen: JsonGenerator, provider: SerializerProvider) =
+                jgen.writeNumber(value.toEpochMilliseconds())
+        }
+
+        class KotlinInstantDeserializer : JsonDeserializer<Instant>() {
+            override fun deserialize(p: JsonParser, p1: DeserializationContext): Instant =
+                Instant.fromEpochMilliseconds((p.codec.readTree(p) as JsonNode).longValue())
+        }
+    }
+}
